Update comments Action Func Delegate for evaluation
